+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- **CUMULUS-4898**
+  - Add Iceberg API documentation page to Cumulus Documentation
+    Once released, the Iceberg API doc should be at: https://nasa.github.io/cumulus/docs/next/deployment/iceberg-api
+- **CUMULUS-4891**
+  - Add scripts to build Iceberg replication docker images and push them to ECR as part of the build process
 - **CUMULUS-4866**
   - Add cmrProvider to sns output granules, executions, pdrs
 - **CUMULUS-4883**
@@ -29,6 +34,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Set max_locks_per_transaction database parameter to 256 to support better performance with new partitioning setup.
 
 ### Changed
+
+- **CUMULUS-4891** Force build/push of iceberg replication images when merging to master
 
 - **CUMULUS-4776** Split iceberg replication into separate services and add support for partitioned tables